+++ test/unit/clientDecorator/plugins/index.js
@@ -1,0 +1,3 @@
+describe('Plugins', function () {
+	require('./plain');
+});