+++ test/unit/index.js
@@ -1,3 +1,4 @@
 describe('Unit Tests', function () {
 	require('./clientDecorator');
+	require('./taskDecorator');
 });