+++ test/unit/taskDecorator/index.js
@@ -1,14 +1,9 @@
 var sinon = require('sinon');
 
-describe('Client-Decorators', function () {
+describe('Task-Decorators', function () {
 
 	beforeEach(function () {
-
 		this.sandbox = sinon.sandbox.create();
-
-		this.eventReporter = {
-			on: this.sandbox.stub()
-		};
 	});
 
 	afterEach(function () {