+++ test/unit/taskDecorator/plugins/index.js
@@ -1,0 +1,5 @@
+describe('Plugins', function () {
+	require('./identifier');
+	require('./decorator');
+	require('./group');
+});