+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.9.8
 * Major logging refactor
 * Deprecate "subprofile" parameter in favor of "sub-profile" for more consistency in parameters
 * Remove "verbose" flag on task
+* Remove "bail" from group-task - use default "failOnError"
 
 v0.9.7 - 12/02/14
 * Updated dependencies